+++ sdk/chaos/arm-chaos/CHANGELOG.md
@@ -1,15 +1,41 @@
 # Release History
+    
+## 1.0.0-beta.4 (2023-03-17)
+    
+**Features**
 
-## 1.0.0-beta.4 (Unreleased)
+  - Added operation Experiments.beginCancel
+  - Added operation Experiments.beginCancelAndWait
+  - Added operation Experiments.beginCreateOrUpdate
+  - Added operation Experiments.beginCreateOrUpdateAndWait
+  - Added operation Experiments.beginStart
+  - Added operation Experiments.beginStartAndWait
+  - Added operation Experiments.update
+  - Added Interface ComponentsEwb5TmSchemasUserassignedidentitiesAdditionalproperties
+  - Added Interface ExperimentsCancelHeaders
+  - Added Interface ExperimentsStartHeaders
+  - Added Interface ExperimentsUpdateOptionalParams
+  - Added Interface ExperimentUpdate
+  - Added Type Alias ExperimentsUpdateResponse
+  - Interface ExperimentsCancelOptionalParams has a new optional parameter resumeFrom
+  - Interface ExperimentsCancelOptionalParams has a new optional parameter updateIntervalInMs
+  - Interface ExperimentsCreateOrUpdateOptionalParams has a new optional parameter resumeFrom
+  - Interface ExperimentsCreateOrUpdateOptionalParams has a new optional parameter updateIntervalInMs
+  - Interface ExperimentsStartOptionalParams has a new optional parameter resumeFrom
+  - Interface ExperimentsStartOptionalParams has a new optional parameter updateIntervalInMs
+  - Interface ResourceIdentity has a new optional parameter userAssignedIdentities
+  - Add parameters of ExperimentsCancelHeaders to TypeAlias ExperimentsCancelResponse
+  - Add parameters of ExperimentsStartHeaders to TypeAlias ExperimentsStartResponse
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Removed operation Experiments.cancel
+  - Removed operation Experiments.createOrUpdate
+  - Removed operation Experiments.start
+  - Interface Selector no longer has parameter filter
+  - Removed Enum KnownFilterType
+    
+    
 ## 1.0.0-beta.3 (2023-01-12)
     
 **Features**